+++ Package.swift
@@ -25,13 +25,13 @@ let package = Package(
     targets: [
         .binaryTarget(
             name: "SendbirdUIKit",
-            url: "https://github.com/sendbird/sendbird-uikit-ios/releases/download/3.33.1/SendbirdUIKit.xcframework.zip", // SendbirdUIKit_URL
-            checksum: "b55c89d32f9634270be305c92fc67c0c0b04c23328d4f9436e55037dce3ec9ca" // SendbirdUIKit_CHECKSUM
+            url: "https://github.com/sendbird/sendbird-uikit-ios/releases/download/3.34.0/SendbirdUIKit.xcframework.zip", // SendbirdUIKit_URL
+            checksum: "fbc41adc410b034454fe2b507a0c6b5bbb42c4f4963c438951c94c0a46c39cca" // SendbirdUIKit_CHECKSUM
         ),
         .binaryTarget(
             name: "SendbirdUIMessageTemplate",
-            url: "https://github.com/sendbird/sendbird-uikit-ios/releases/download/3.33.1/SendbirdUIMessageTemplate.xcframework.zip", // SendbirdUIMessageTemplate_URL
-            checksum: "99be28b7f9e23a5749f28565362a335e709d2752fada6af15032879882f69542" // SendbirdUIMessageTemplate_CHECKSUM
+            url: "https://github.com/sendbird/sendbird-uikit-ios/releases/download/3.34.0/SendbirdUIMessageTemplate.xcframework.zip", // SendbirdUIMessageTemplate_URL
+            checksum: "8fb8dd6c05ce84eab5754ed91da7f5ef5e104d60e39ef75c4f995cc618b693f4" // SendbirdUIMessageTemplate_CHECKSUM
         ),
         .target(
             name: "SendbirdUIKitTarget",